+++ packages/elements/src/core/bind/index.ts
@@ -2,7 +2,7 @@ import updateStateAttributes from "./update-state-attributes.js";
 import registerActionEffects from "./register-action-effects.js";
 
 const bind = {
-	updateAttributes: updateStateAttributes,
+	updateStateAttributes,
 	registerActionEffects,
 };