+++ packages/elements/src/core/effect/index.ts
@@ -1,0 +1,7 @@
+import registerEffects from "./register-effects.js";
+
+const effect = {
+	registerEffects: registerEffects,
+};
+
+export default effect;